Internal: Move popover action (and floating actions) [ED-22647]
