+++ packages/packages/core/editor-editing-panel/src/controls-actions.ts
@@ -1,7 +1,7 @@
+import { PopoverAction } from '@elementor/editor-ui';
 import { createMenu } from '@elementor/menus';
 
 import Action from './action';
-import { PopoverAction } from './popover-action';
 
 export const controlActionsMenu = createMenu( {
 	components: {
