+++ packages/packages/libs/editor-ui/src/components/floating-bar.tsx
