+++ packages/packages/libs/editor-ui/src/components/popover/index.ts
@@ -3,3 +3,4 @@ export { SectionPopoverBody } from './section-popover-body';
 export { PopoverHeader } from './header';
 export { ITEM_HEIGHT, PopoverMenuList, StyledMenuList } from './menu-list';
 export type { PopoverMenuListProps, VirtualizedItem } from './menu-list';
+export { PopoverAction, type PopoverActionProps } from './popover-action';
